add read and write scale scripts.
The scripts are made async.  This requires that they are awaited for the output.  In order for all plugins to use the scrips device-utils.js functions tagValueCompose (for read) and tagRawCalculator (for write) were modified to call the scale method.  These were then made async.  This bubbled up to all the plugins.  Requiring changes in every plugin.

Also allowed the user to define the script with additional parameters.  The values for these parameters are set in the tag option panel.  This way the same script can be used for multiple tags, with some tag specific data being passed to the script.  Also the same script can be used for read and write but have different parameter values passed to it.